Ignore the .user file generated by Visual Studio.
Add in a SiteIndexTemplate class which uses a separate template (siteindex.html) to generate the index.html file for the root folder. This is so that I can have custom content on my homepage which does not fall within the structure of a page (article title, etc) or blog post.

Also prevent index.html generation if there is a static index.html file present - in that case, just copy it over untouched.