Remove MarkdownSharp project - it doesn't handle generating code blocks well (e.g. ``` -> <code>).
Replaced with a submodule reference to MarkdownSharp-GithubCodeBlocks (https://github.com/KyleGobel/MarkdownSharp-GithubCodeBlocks/) which surrounds the generated <code> tag with a <pre> tag, thereby preserving line breaks. This is just supplied as a .cs file, so add that to the Scatter project as a linked file.

Remove the getMarkdownSource.sh file since I'm no longer using that version of MarkdownSharp.